caso2_objetual/requirements.txt
psycopg2-binary==2.9.7
python-dotenv==1.0.0
